Update USER_STORIES.md
I found the problem I was having with the anchor links. Some of the emojis are complex, and seem to be emojis with modifiers or something similar. When I delete the "top layer," it leaves behind a ghost emoji. It works perfectly fine when using the ghost emoji, but does not work when using the full emoji. I deleted the originals and the links appear to be working. I'll test and return if something is incorrect.