So a few things on this commit/update:
wanted to wrap the pages with MainContainer, for a simple page layout style. However, I also wanted to wrap the pages with the Header and Footer. So I created a separate PageLayout.jsx file to act as a hub for all the elements and components that would make up the majority of pages.

Once I created that, I simple removed the Header and style parts of the WelcomePage.jsx and LoginPage.jsx code and replaced it with the PageLayout wrapper.

After tweaking a couple bugs and figuring a couple things out, I've tested it via npm on my local and it works!

Ultimately, everything I've done hasn't changed the visual of the website at all, but it has avoided DRY and saved some work in the future, so horray!