Added import for React to LoginPage.jsx
Added import for LoginPage to App.jsx. Also added the LoginPage route to App.jsx

Added code to StartQuestButton.jsx to create a simple button to navigate to the login page. Used the routes to route to the LoginPage directly from the button.

Placed the button on WelcomePage and included the necessary import.

Tested it via npm locally and it worked well. LoginPage is successfully rendering with the Header, Footer, and the appropriate body content.