Created a route in App.jsx to lead from LoginPage to Signup Page if a user doesn't have an account.
Added an import to LoginPage for link, and created a link to navigate to SignupPage.

I don't LOVE the extra button presses required from a user, so I may eventually just skip the WelcomePage middle man and jump to login when a user opens up the app, but for now this works okay.