Created a simple typography.js file in styles to house the simple link styling I had been using between LoginPage and SignupPage. At first I thought about just adding it to forms.js, because right now it's only used on form pages to navigate between them, but I want to re-use the styling throughout for other simple links, so for now I'll put it in typography.
There was a bit of an internal debate on whether or not to put it in a layout.js instead, since it involves mostly spacing, but it's possible this simple styling involves a lot more font based stuff, and I'm not 100% sure how I'll go about that. For now, this is fine, but may be refactored into a layout.js in the future.