Created a PrivacyPolicy page with basic skeleton code with the necessary imports for basic functionality.
Added a lastModifiedDate const to PrivacyPage to allow for easy handling when the policy is revised in the future. After creating it as a string, decided to use the built in Date() object, just in case we want to get full access to querying parts of it in the future, if needed.

Added the routes to App.jsx and added the link to the page via the Footer.jsx.

After a bit of testing and debugging, works well. All pages are connected.