Created a How it Works page and placed it in the Header.
Added the corresponding routes to App.jsx

Added the page links to the Header.jsx.

Tested everything, and it works so far. All pages are connected correctly.

I'm not sure how I feel about the right-justification of the elements, however. I'll fill it out first and then make design decisions, but I'm assuming I'll need to wrap it all up into a div style or something so the logo and nav play nice with each other instead of against each other as separate elements.