Created a layout.js file to house global structural elements. Added the nav styling I've been using for footer/header in there, with plans to avoid violating DRY and re-use across navigation elements. Bare minimum, the shared elements should work together nicely in there.
Replaced the code in Header() nav styling to the abstract navContainer.

Tested and it works the same as it did before. Made sure to test correctly by changing elements of navContainer, which affected the page. All working as intended.

Next step is to do the same with Footer and replace that content.