Added a .prettierrc file, to properly format each save file to a standard convention without too much worry. Connects with Prettier and automatically formats files for consistency when they are saved.
Went through all of the files in the project and saved them to add the changes.

Did some light testing to make sure everything still looked and worked the same. Will double-check the documentation after I commit this to make sure it still visually looks good. If not, I may need to put in some kind of ignore for .md files.