Noticed that the Footer wasn't anchored to the bottom of the viewport, and would instead float if there wasn't much content.
Did two things. I added a div and some styling to PageLayout to clean it up a bit and ensure vertical stacking.

Then I added a flexGrow: 1 to MainContainer, so all pages will share this behavior without much issue.

If I end up needing to re-use PageLayout and this is a problem, I can add it to PageLayout with props or something, but this is so much cleaner. I really doubt I'll re-use PageLayout for future cards and such anyway, but it's something to keep in mind for the future.