Just went through hell and back trying to debug why I have a scrollbar in my browser window, despite the content not overflowing. Did a lot of debugging in various places, and changed a few things (but was careful to change them back). I changed a couple margin/padding details in Footer, MainContainer, and PageLayout, but nothing major.
What I found is that my browser was adding 8 px to both the top and bottom margins of the screen. At first I thought it was an extra 16 pixels hidden somewhere (since my browser height was 1313 and my body height was 1297), but it was in body defaults.

Added a style margin to main.jsx to set the default margins.