Added const for heading text style and body text style in styles/typography.
Imported the heading text style to all pages and added it to each h1 on every page.

I may refactor in the near future with some kind of heading function to avoid violating DRY, since I don't LOVE that every page requires me to have h1 style, but idk. That's for another day.