Two commits:
Realized that the SignupPage was missing an import for the heading style, so there was a broken link. Fixed and fixed.

Created a context folder and placed AuthProvider.jsx inside of it. This is to set up React Context to track a logged in user through the app.

I added some basic code to create authentication context. Still need to wire and manage state. Just going a bit at a time.