Updated Changelog and READMe with changes.
The readme is currently a bit of a mess. I'll need to go through and reorganize it and update it completely, with a table of contents, at some point. Eventually it'll look entirely different as well, because right now I would say it's not very user friendly at all. That being said, it works for me for now.