Updated SignupPage.jsx to attach sign up to firebase. Included imports for create user with email and password, and updating profile with information.
I added constants for error message and a success message, and added a try/catch block for error handling.

Lastly, added a bit of styling to that. That styling will eventually be moved as I continue to work on this, but for now it can live on this page.

Tested that sign up worked correctly. After handling a bug (a missing closing bracket), it worked correctly. I created an account, and then deleted the account via the firebase console without issue, and got the correct success message. More testing is needed to confirm various errors, but for now it's working as intended.