Added code and logic to LoginPage to function similarly to SignupPage. It navigates users to Dashboard after a successful sign in.
Right now, there is no way to sign in using a Display Name, and right now Display Name is being auto set to a blank string. So I will need to make sure that it grabs the display name correctly from firebase, even if a user doesn't use one to log in.