Added some logic that toggles the Sign In link to a Log Out button when a user is signed in. With some light testing using my created user, it worked well.
Tested it by navigating to every page, making sure that the user stays logged in, and that the log out button works regardless of where the user navigates.

I still want some kind of confirmation that the user wants to sign out, like a little pop up. I also need to make it so the Trash Quest logo returns them to the dashboard, instead of the welcome page.