Added a simple pop up that confirms with the user if they want to log out or not when they hit the log out button.
Tested and it's working well. The only issue is that it also pops up when the user signs in, which is awkward. It doesn't break the app or really affect functionality, so I'm going to push this now and then take a look and see what I can figure out.