Added Features section to README, and cleaned up README in general. Added more sections to ToC and cleaned up the formatting so it looks a lot better.
It's still a mess to be honest, and any non-tech user looking at this would be a challenge, but for now it's working well to at least document everything that the app is doing, so I'm going to keep it. Eventually I want to clean it up and maybe separate this out so Contributers can see most of this, but regular users will just see the typical App features and so on.