Created my second test and finished adding the code!
Test -> redirects to default route if unathenticated

In other words, if the user is not logged in, protected content should not be available and it should redirect the user.

The second test was much easier, since I had already ironed out the syntax and weird file-level bugs.

Tested the test to make sure that it was correctly testing the right thing. Everything is passing as expected!