Connected the basic forgot password modal to Login page via a button.
As of right now, it's unstyled and quite basic, so it just appears in kind of an ugly way at the bottom of the screen. It also does not actaully provide a way for user to use the feature just yet, as this was simply to see if the modal would work at all. So far, it's working!