Updated ForgotPasswordModal to add actual firebase logic and functionality. Now it gives a form response that allows users to type in their email.
Connects to all the usual functionality, including FormButton and FormGroup.

I tested, and I was able to get a forgot password reset email in my inbox. The email is a bit ugly/unformatted, so I'll need to figure out how to customize that through firebase, but so far so good.

On top of that, it did default to my SPAM folder, which is understandable, but that's something else for me to consider as I move forward. The email link also doesn't work. I'm running into issues where the old api key is being used for things like that, so I contacted support. Tried to debug for hours with no fix in sight.