Added 8th (and I think last, but we'll see) test to FormGroup.
Test 8: Verifies that OnChange is properly called when the user types into a FormGroup field.

Uses fireEvent to simulate user behavior for typing. When the user types, onChange should trigger exactly once.

This guarantees that it isn't being called a bunch of times, and it also tests to make sure it's being called at all.