Refactored a couple of things.
First, I realized that my test 4 was using newly created mocks that are redundant to the props already set up as default at the top of the describe block, so I fixed that.

After a bunch of debugging, I found that the test was now failing, but I figured out that I had two mock functions that weren't connected. Connected them, and now everything is running smoothly.