Completed the ConfirmLoadout testing suite, at least for now! Added 4 more tests:
Test 7: Verifies that "Yes" Button is indeed a button component

Test 8: Verifies that "No" Button is indeed a button component

Test 9: Verifies that onConfirm is not called on render

Test 10: Verifies that onCancel is not called on render

All tests are passing!

Also created Loader.test.jsx and added it to the scaffold. Nothing was added to the file just yet.