Haha what a journey! I went down a crazy rabbit hole when I tried to add a 7th test for Modal.test. Basically, I was checking if the modal would close properly if clicked outside of the modal area. This made me realize that functionality wasn't actually present, as there was no styling. Because of this, I started to add styling, which made me dive into a variety of issues and visual problems with overlay.
After a ton of debugging, I realized that I was trying to use modalContentStyle inside of a button instead of the proper div tag, so that created all sorts of issues.

After a bunch of work, I finally have the modal working correctly! When it pops up, it grays out the background, and a clear modal with a white background and options appears. The text appears correctly, the error codes appear correctly, and the various fields and buttons appear correctly.

I did need to call the modal styles directly in ForgotPassword to allow for separate styling for different text styles, like h2 and p. I'm not 100% convinced this is the best approach instead of somehow adding it directly into the Modal.jsx file, but it works really well for now and I don't want to get too crazy.