Added the last remaining tests for Footer, and reorganized tests based on function/link.
Basically, I added a test to verify that the link elements are inside of the nav elements, and not moved or placed elsewhere. This test isn't the MOST helpful, but it verifies that these links aren't for some reason moved outside of nav on accident.

I created one of these tests for all 4 elements, and grouped them by link text name.

All 18 tests are passing!