Changed/deleted some of my submissions scaffold that I designed over a month ago. Now fitting my new design, which does not include photo uploads for a plethora of reasons (security, privacy, accessibility, and user participation).
This type select will be used in the LogCleanupForm.

Still need to create a test file for it, and to actually add it to a greater Log submission form, so I'm sure I'll run into something weird when that happens. For now, I'll save it and start work on LogCleanup.