Added a LogCleanupForm, which declares eco points values and clean up types, and area options.
From there, it should ask the user for some information for their trash clean up, and based on that information, will calculate basic trash clean up values and show them.

If a user clicks Hazardous Waste, a warning text will pop up, letting them know to go by state and city laws, as well as seek help from a professional if needed.

This component is still untested. I'll create tester files for it and TrashTypeSelect, but first I want to just add some simple code to LogCleanupPage and add the form to it, and see what happens. From there, if it's at least working, I'll add tests to help debug/validate that everything is working as intended.

I also added some TODOs for future feature, like saving and updating user information based on the logs, as I get this up and running.