Changed file name from LogCleanup to LogCleanupPage to match the rest of my pages.  From there, added basic code that matches the rest of my pages, to just have a brief user welcome to the log cleanup page.
I also added the route to App.jsx.

I still need to add it to Header, where it will be selective based on user authorization (they will only see it when they are logged in).

After that, I will test nav flow, and then once everything is working, will add in the LogCleanupForm to the main page and do manual testing before I create a testing file for the component.