Added a protected route to log-cleanup so it can't be navigated to unless the user is logged in. Otherwise, it just renavigates to the login page.
However, I did notice a major inconsistency with how I'm handling both ProtectedRoute and UseRedirect. They are in different locations, and seemingly handle the same thing, different ways. I'm going to standardize this and hopefully not break everything lol.

In the meantime, I've tested that an unauthenticated user navigating to /log-cleanup is sent to the login page, and it works :)