Created a CleanupEntriesList.jsx file in components\eco. It's job is to show the cleanup entries list for the Cleanup Entries Page and the DashboardPage. It takes in props to show a view button and to limit the list to the last 3 entries. In the Dashboard, it will only show the latest 3 entries and then provide a button to go to an entries page where all will be shown (that page hasn't been created yet, however, so currently this button.... doesn't work lol).
I'm going to do that next, however! For now, I just added it to my Dashboard to make sure that part was working correctly. I'm also going to refactor this and piece out a lot of the component/reusable parts. I just like to work with single file first and then split it up/use reusables as I can.

I also had to add an index to Firebase in order to do order by and whatnot, for cleanup entries.

For now, Dashboard works well, but the view button does not.