Created a CleanupEntriesListPage and added simple boilerplate to it so that it would work okay as a page. This page WILL take on the CleanupEntriesList component and show the entire list, but I like to start with the simple page first and go from there.
I added the route to App.jsx for it, and manually tested that all routes and everything is working. It is!

Next is to add the logic to the page to show actual cleanup entries for the user! It should be paginated in some way, so I'll figure that out. I"ll need to create test entries to test pagination. We'll see how that goes lol. I'll probably just add the entries first and worry about pagination later.