Updated CleanupEntriesListPage to limit the number of entries that first pop up. It offers a show all entries button, which uses the reusable button styling for primary and medium size.
I was planning on doing pagination, but it is a bit more complicated/involved than I was anticipating, and I don't want to get too trapped in the rabbit hole at the moment. I added it to my Roadmap and updated my roadmap in general!

I didn't change EcoBadge at all so I have no idea why it's saying I made changes. Even the change itself is just an add space that is already part of my auto-formatting, so now clue what it's on about. Whatever lol.