Made a few changes.
added a isCancel prop to FormButton to allow for differences in cancel button styling. Added an if statement that checks if the button isCancel or not. If it is, it sets the buttonColor to secondary (will likely change this as I add more colors to the app).

Passed that prop into the FormButton call in LogCleanupForm for the cancel button, so it's used properly.

Deleted formButtonStyle from forms.js, with the plan to just use buttonStyles for all buttons in the app.

Lastly, I deleted a couple things from my jsconfig.json file because it was assuming all of my imports are at the base level, and it was annoying me getting wrong imports constantly.