Added delete logic to allow the user to delete their cleanup entries.
However, I still need it to update the firestore for user points total reversal and all that. I also need to add a popup that asks the user if they are sure before deletion.

In addition, we need to figure out the best way to handle malicious use (someone deleting all entries, for example), and how to handle potential negative points totals (points were spent, and then the entry that provided those points is deleted).