Added a DeleteCleanupEntryModal file that creates a modal when a user tries to delete a log entry. It will give them information about the entry they want to delete, and how it will affect their current profile.
Asks for a confirmation, and on confirmation it updates user's points and so on.

Added that logic to my CleanupEntriesList.

However, I don't like that this responsibility of data fetching and updating is split between both files. I'm going to create a hook and move a bunch of stuff into it so this is cleaned up, and the parent (CleanupEntriesList) is actually directly responsible for all of this, instead of its child modal component.

Currently, it always shows current user's points as 0. I plan on changing this when I create the hook.