Abstracted out some utility style styles as well as some general wrappers for the cleanup entries list, and referenced them correctly.
As I continue to work on this, I'm sure I'll find some re-used styles and rename them as needed, but for now this is working well.

Also got rid of what seems to be a redundant div style and just combined two together. In the future I may need to separate it, but it works great now.

entriesListLayoutWrapper
centerbuttonLayout
containerElementSpacing
entryContainerWrapper